+++ src/tree/inner_node.rs
@@ -1,4 +1,4 @@
-use crate::*;
+use super::*;
 use std::alloc::alloc;
 use std::slice::SliceIndex;
 use std::{
