+++ src/tree/leaf_node.rs
@@ -1,4 +1,4 @@
-use crate::*;
+use super::*;
 use std::{
     alloc::{alloc, Layout},
     mem::{self, MaybeUninit},
